+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release notes
 
-## Unreleased
+## 0.3.55 — 2026-09-11
 
 ### Changed